Revert "Auth test 2"
