+++ client/package.json
@@ -28,7 +28,6 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-  "proxy": "http://localhost:1234",
   "eslintConfig": {
     "extends": "react-app"
   },
