+++ client/src/components/containers/singleUserContainer.js
@@ -1,29 +1,57 @@
 import React, { Component } from 'react';
 import SingleUserView from '../views/singleUserView';
 import { connect } from 'react-redux';
-import { withRouter } from "react-router-dom";
-import { fetchSessionsThunk , addStudySessionThunk , currStudySessionThunk, logout } from '../../store/actions/actionCreatorsThunks';
+import { fetchSessionsThunk , addStudySessionThunk , currStudySessionThunk } from '../../store/actions/actionCreatorsThunks';
 
 // import AppBar from '@material-ui/core/AppBar'
 // import Toolbar from '@material-ui/core/Toolbar'
 // import Typography from '@material-ui/core/Typography'
 
 class SingleUserContainer extends Component {
+    // constructor(props){
+    //     super(props)
+    //     this.state = {
+    //         videoUrl: "",
+    //         studySessionName: "",
+    //         studySessionDescription: "",
+    //         userId: 0
+    //     }
+    // }
+
+    // handleChange=(event)=>{
+    //     this.setState({
+    //         [event.target.name]: event.target.value
+    //     })
+    // }
+
+    // handleSubmit=(event)=>{
+    //     event.preventDefault();
+    //     let study_sess = {
+    //         videoUrl: this.state.videoUrl,
+    //         studySessionName: this.state.studySessionName,
+    //         studySessionDescription: this.state.studySessionDescription,
+    //         userId: this.state.userId
+    //     }
+    //     this.props.addStudySessionThunk(study_sess);
+    // }
+
     componentDidMount() {
         this.props.fetchSessionsThunk();
-    }
-    //auth
-    handleLogout = () => {
-        this.props.logout();
-        this.props.history.push("/login");
     }
     render(){
         return(
             <div>
+                {/* <AppBar position="static">
+                    <Toolbar>
+                        <Typography variant="title" color="inherit">
+                        React & Material-UI Sample Application
+                        </Typography>
+                    </Toolbar>
+                </AppBar> */}
 
-
-                <SingleUserView sessions = {this.props.userSessions} handleLogout={this.handleLogout} 
-
+                <SingleUserView sessions = {this.props.userSessions}
+                    // handleChange = {this.handleChange} 
+                    // handleSubmit={this.handleSubmit}
                     currentStudySession={this.props.currStudySessionThunk}
                 />
             </div>
@@ -41,9 +69,8 @@ const mapDispatch = (dispatch) => {
     return({
         fetchSessionsThunk: () => dispatch(fetchSessionsThunk()),
         // addStudySessionThunk: (study_sess) => dispatch(addStudySessionThunk(study_sess)),
-        currStudySessionThunk: (study_sess) => dispatch(currStudySessionThunk(study_sess)),
-        logout: () => dispatch(logout())
+        currStudySessionThunk: (study_sess) => dispatch(currStudySessionThunk(study_sess))
     })
 }
 
-export default withRouter(connect(mapState, mapDispatch)(SingleUserContainer));
+export default connect(mapState, mapDispatch)(SingleUserContainer);
