+++ client/src/store/actions/actionTypes.js
@@ -13,8 +13,3 @@ export const CURRENT_VIDEO = 'CURRENT_VIDEO';
 // export const REGISTER_SUCCESS = 'REGISTER_SUCCESS';
 // export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
 // export const LOG_OUT = 'LOG_OUT';
-
-//AUTH
-export const GET_USER = "GET_USER";
-export const REMOVE_USER = "REMOVE_USER";
-
