--- client/src/store/reducers/userAuth.js
@@ -1,13 +1,0 @@
-import * as type from '../actions/actionTypes';
-
-
-export default(state = {}, action) =>{
-    switch(action.type){
-        case (type.GET_USER):
-            return (action.payload);
-        case (type.REMOVE_USER):
-            return {};
-        default:
-            return state; 
-    }
-}
