+++ server/database/models/User.js
@@ -1,5 +1,4 @@
 //	This is how the table of User is set up
-const crypto = require("crypto");
 const Sequelize = require('sequelize');
 const db = require('../db');
 
@@ -11,53 +10,25 @@ const User = db.define("user", {
 		autoIncrement: true
 	},
 
-	username: {
+	userName: {
 		type: Sequelize.STRING,
-		unique: true,
 		allowNull: false
 	},
 
-	password: {
+	pass: {
+		type: Sequelize.TEXT,
+		allowNull: false,
+	},
+
+	email: {
 		type: Sequelize.STRING,
 		allowNull: false,
-		get(){
-			return () => this.getDataValue("password");
+		validate: {
+			notEmpty: true,
+			isEmail: true
 		}
 	},
 
-	salt: {
-		type: Sequelize.STRING,
-		get() {
-			return () => this.getDataValue("salt");
-		}
-	}
 });
-
-User.generateSalt = function() {
-	return crypto.randomBytes(16).toString("base64");
-};
-
-User.encryptPassword = function(plainText, salt){
-	return crypto
-		.createHash("RSA-SHA256")
-		.update(plainText)
-		.update(salt)
-		.digest("hex");
-};
-
-User.prototype.correctPassword = function(passwordAttempt) {
-	return User.encryptPassword(passwordAttempt, this.salt()) === this.password();
-
-};
-
-const setSaltAndPassword = user => {
-	if (user.changed("password")) {
-		user.salt = User.generateSalt();
-		user.password = User.encryptPassword(user.password(), user.salt());
-	}
-};
-
-User.beforeCreate(setSaltAndPassword);
-User.beforeUpdate(setSaltAndPassword);
 
 module.exports = User;
