+++ server/package.json
@@ -7,15 +7,12 @@
   },
   "dependencies": {
     "compression": "^1.7.3",
-    "connect-session-sequelize": "^6.0.0",
     "cookie-parser": "~1.4.3",
     "cors": "^2.8.5",
     "debug": "~2.6.9",
     "express": "~4.16.0",
-    "express-session": "^1.17.0",
     "helmet": "^3.14.0",
     "morgan": "~1.9.0",
-    "passport": "^0.4.1",
     "pgtools": "^0.3.0",
     "sequelize": "^5.15.1"
   }
